Move all routes into subfolder
Move all routes into a subfolder 'routes' to modularise code.
