+++ routes/routes_encounter.py
